requirements.txt
mysql-connector-python
python-dotenv
google-generativeai